requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.15.2
psycopg2-binary==2.9.10
sqlparse==0.5.3
python-dotenv==0.21
resend==2.4.0
django-cors-headers==4.6.0